+++ src/resources/drafts.ts
@@ -114,12 +114,28 @@ export class Drafts extends Resource {
     requestBody,
     overrides,
   }: CreateDraftParams & Overrides): Promise<NylasResponse<Draft>> {
-    const form = Messages._buildFormRequest(requestBody);
+    const path = `/v3/grants/${identifier}/drafts`;
 
-    return this.apiClient.request({
-      method: 'POST',
-      path: `/v3/grants/${identifier}/drafts`,
-      form,
+    // Use form data only if the attachment size is greater than 3mb
+    const attachmentSize =
+      requestBody.attachments?.reduce(function(_, attachment) {
+        return attachment.size || 0;
+      }, 0) || 0;
+
+    if (attachmentSize >= Messages.MAXIMUM_JSON_ATTACHMENT_SIZE) {
+      const form = Messages._buildFormRequest(requestBody);
+
+      return this.apiClient.request({
+        method: 'POST',
+        path,
+        form,
+        overrides,
+      });
+    }
+
+    return super._create({
+      path,
+      requestBody,
       overrides,
     });
   }
@@ -134,12 +150,28 @@ export class Drafts extends Resource {
     requestBody,
     overrides,
   }: UpdateDraftParams & Overrides): Promise<NylasResponse<Draft>> {
-    const form = Messages._buildFormRequest(requestBody);
+    const path = `/v3/grants/${identifier}/drafts/${draftId}`;
 
-    return this.apiClient.request({
-      method: 'PUT',
-      path: `/v3/grants/${identifier}/drafts/${draftId}`,
-      form,
+    // Use form data only if the attachment size is greater than 3mb
+    const attachmentSize =
+      requestBody.attachments?.reduce(function(_, attachment) {
+        return attachment.size || 0;
+      }, 0) || 0;
+
+    if (attachmentSize >= Messages.MAXIMUM_JSON_ATTACHMENT_SIZE) {
+      const form = Messages._buildFormRequest(requestBody);
+
+      return this.apiClient.request({
+        method: 'PUT',
+        path,
+        form,
+        overrides,
+      });
+    }
+
+    return super._update({
+      path,
+      requestBody,
       overrides,
     });
   }